+++ lib/capybara/spec/session/first_spec.rb
@@ -9,8 +9,14 @@ Capybara::SpecHelper.spec '#first' do
     expect(@session.first("//input[@id='test_field']").value).to eq('monkey')
   end
 
-  it "should return nil when nothing was found" do
-    expect(@session.first('//div[@id="nosuchthing"]')).to be_nil
+  it "should raise ElementNotFound when nothing was found" do
+    expect do
+      @session.first('//div[@id="nosuchthing"]')
+    end.to raise_error Capybara::ElementNotFound
+  end
+
+  it "should return nil when nothing was found if count options allow no results" do
+    expect(@session.first('//div[@id="nosuchthing"]', minimum: 0)).to be_nil
   end
 
   it "should accept an XPath instance" do
@@ -22,7 +28,7 @@ Capybara::SpecHelper.spec '#first' do
 
   it "should warn when unused parameters are passed" do
     expect_any_instance_of(Kernel).to receive(:warn).with(/Unused parameters passed.*unused text/)
-    @session.first(:css, '.header h2', 'unused text')
+    @session.first(:css, 'h1', 'unused text')
   end
 
   context "with css selectors" do
@@ -49,40 +55,52 @@ Capybara::SpecHelper.spec '#first' do
 
   context "with visible filter" do
     it "should only find visible nodes when true" do
-      expect(@session.first(:css, "a#invisible", visible: true)).to be_nil
+      expect do
+        @session.first(:css, "a#invisible", visible: true)
+      end.to raise_error Capybara::ElementNotFound
     end
 
     it "should find nodes regardless of whether they are invisible when false" do
-      expect(@session.first(:css, "a#invisible", visible: false)).not_to be_nil
-      expect(@session.first(:css, "a#invisible", visible: false, text: 'hidden link')).not_to be_nil
-      expect(@session.first(:css, "a#visible", visible: false)).not_to be_nil
+      expect(@session.first(:css, "a#invisible", visible: false)).to be
+      expect(@session.first(:css, "a#invisible", visible: false, text: 'hidden link')).to be
+      expect(@session.first(:css, "a#visible", visible: false)).to be
     end
 
     it "should find nodes regardless of whether they are invisible when :all" do
-      expect(@session.first(:css, "a#invisible", visible: :all)).not_to be_nil
-      expect(@session.first(:css, "a#invisible", visible: :all, text: 'hidden link')).not_to be_nil
-      expect(@session.first(:css, "a#visible", visible: :all)).not_to be_nil
+      expect(@session.first(:css, "a#invisible", visible: :all)).to be
+      expect(@session.first(:css, "a#invisible", visible: :all, text: 'hidden link')).to be
+      expect(@session.first(:css, "a#visible", visible: :all)).to be
     end
 
     it "should find only hidden nodes when :hidden" do
-      expect(@session.first(:css, "a#invisible", visible: :hidden)).not_to be_nil
-      expect(@session.first(:css, "a#invisible", visible: :hidden, text: 'hidden link')).not_to be_nil
-      expect(@session.first(:css, "a#invisible", visible: :hidden, text: 'not hidden link')).to be_nil
-      expect(@session.first(:css, "a#visible", visible: :hidden)).to be_nil
+      expect(@session.first(:css, "a#invisible", visible: :hidden)).to be
+      expect(@session.first(:css, "a#invisible", visible: :hidden, text: 'hidden link')).to be
+      expect do
+        @session.first(:css, "a#invisible", visible: :hidden, text: 'not hidden link')
+      end.to raise_error Capybara::ElementNotFound
+      expect do
+        @session.first(:css, "a#visible", visible: :hidden)
+      end.to raise_error Capybara::ElementNotFound
     end
 
     it "should find only visible nodes when :visible" do
-      expect(@session.first(:css, "a#invisible", visible: :visible)).to be_nil
-      expect(@session.first(:css, "a#invisible", visible: :visible, text: 'hidden link')).to be_nil
-      expect(@session.first(:css, "a#visible", visible: :visible)).not_to be_nil
+      expect do
+        @session.first(:css, "a#invisible", visible: :visible)
+      end.to raise_error Capybara::ElementNotFound
+      expect do
+        @session.first(:css, "a#invisible", visible: :visible, text: 'hidden link')
+      end.to raise_error Capybara::ElementNotFound
+      expect(@session.first(:css, "a#visible", visible: :visible)).to be
     end
 
     it "should default to Capybara.ignore_hidden_elements" do
       Capybara.ignore_hidden_elements = true
-      expect(@session.first(:css, "a#invisible")).to be_nil
+      expect do
+        @session.first(:css, "a#invisible")
+      end.to raise_error Capybara::ElementNotFound
       Capybara.ignore_hidden_elements = false
-      expect(@session.first(:css, "a#invisible")).not_to be_nil
-      expect(@session.first(:css, "a")).not_to be_nil
+      expect(@session.first(:css, "a#invisible")).to be
+      expect(@session.first(:css, "a")).to be
     end
   end
 
@@ -93,35 +111,34 @@ Capybara::SpecHelper.spec '#first' do
 
     it "should find the first element using the given locator" do
       @session.within(:xpath, "//div[@id='for_bar']") do
-        expect(@session.first('.//form')).not_to be_nil
+        expect(@session.first('.//form')).to be
       end
     end
   end
 
-  context "with Capybara.wait_on_first_by_default", requires: [:js] do
+  context "waiting behavior", requires: [:js] do
     before do
       @session.visit('/with_js')
     end
 
-    it "should not wait if false" do
-      Capybara.wait_on_first_by_default = false
+    it "should not wait if minimum: 0" do
       @session.click_link('clickable')
-      expect(@session.first(:css, 'a#has-been-clicked')).to be_nil
+      expect(@session.first(:css, 'a#has-been-clicked', minimum: 0)).to be_nil
     end
 
-    it "should wait for at least one match if true" do
-      Capybara.wait_on_first_by_default = true
+    it "should wait for at least one match by default" do
       Capybara.using_wait_time(3) do
         @session.click_link('clickable')
         expect(@session.first(:css, 'a#has-been-clicked')).not_to be_nil
       end
     end
 
-    it "should return nil after waiting if no match" do
-      Capybara.wait_on_first_by_default = true
+    it "should return nil after waiting if no match and allow_nil is true" do
+      @session.click_link('clickable')
+      start_time = Time.now
       Capybara.using_wait_time(3) do
-        @session.click_link('clickable')
-        expect(@session.first(:css, 'a#not-a-real-link')).to be_nil
+        expect(@session.first(:css, 'a#not-a-real-link', allow_nil: true)).to be_nil
+        expect(Time.now-start_time).to be > 3
       end
     end
   end