+++ spec/shared_selenium_session.rb
@@ -57,16 +57,16 @@ RSpec.shared_examples "Capybara::Session" do |session, mode|
     describe "#accept_alert" do
       it "supports a blockless mode" do
         @session.visit('/with_js')
-        skip "Headless Chrome doesn't support blockless modal methods" if @session.driver.send(:headless_chrome?)
         @session.click_link('Open alert')
         @session.accept_alert
         expect{@session.driver.browser.switch_to.alert}.to raise_error(@session.driver.send(:modal_error))
       end
 
-      it "raises if block is missing" do
-        @session.visit('/with_js')
-        skip "Only Headless Chrome requires the block due to system modal JS injection" unless @session.driver.send(:headless_chrome?)
-        expect { @session.accept_alert }.to raise_error(ArgumentError)
+      it "can be called before visiting" do
+        @session.accept_alert "Initial alert" do
+          @session.visit('/initial_alert')
+        end
+        expect(@session).to have_text('Initial alert page')
       end
     end