+++ History.md
@@ -15,6 +15,13 @@ Release date: unreleased
 
 * `Capybara.exact_options` no longer exists. Just use `exact:true` on relevant actions/finders if necessary.
 
+#Version 2.18.0
+Release date: unreleased
+
+### Removed
+
+* Headless chrome modal JS injection that is no longer needed for Chrome 64+/chromedriver 2.35+
+
 
 # Version 2.17.0
 Release date: 2018-01-02